Create Logical Statements
You can win cash by playing the lottery. If the ticket number is a multiple of 10 and 9, the program outputs “You won $200”.

If it is a multiple of 4 or 6, the program outputs “You won $50”. In all other cases, there is no prize and the output is “Try again”.

Sample Input
11457

Sample Output
Try again

Hint
Number a is a multiple of number b, if a%b == 0.